pyproject.toml
[project]
name = "compression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "av>=15.1.0",
    "ffmpeg-python>=0.2.0",
    "icecream>=2.1.7",
    "numpy>=2.3.2",
    "torchinfo>=1.8.0",
]
